+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+-   `ListItem`: Added `isDisabled` prop to disable a clickable list item.
+
 ## [2.1.2][] - 2021-11-10
 
 ### Fixed