+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.1.3][] - 2021-11-19
+
 ### Added
 
 -   `ListItem`: Added `isDisabled` prop to disable a clickable list item.
@@ -1278,3 +1280,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [2.1.1]: https://github.com/lumapps/design-system/tree/v2.1.1
 [unreleased]: https://github.com/lumapps/design-system/compare/v2.1.2...HEAD
 [2.1.2]: https://github.com/lumapps/design-system/tree/v2.1.2
+[unreleased]: https://github.com/lumapps/design-system/compare/v2.1.3...HEAD
+[2.1.3]: https://github.com/lumapps/design-system/tree/v2.1.3